tests: DB_Column_Meta_Compat trait handles int widths + Extra 'NULL' across MySQL/MariaDB
