+++ README.md
@@ -217,6 +217,51 @@ try{
 ### MIT License ###
 http://www.opensource.org/licenses/mit-license.html  
 
+<details markdown="1">
+<summary><strong>All 200 tested combinations</strong> — 5 PHP × 4 WordPress × 10 database — click to expand</summary>
+
+### WordPress 6.6
+
+| PHP \ DB | mysql:9 | mysql:8.4 | mysql:8.0 | mariadb:12 | mariadb:11.8 | mariadb:11.4 | mariadb:11.2 | mariadb:11.0 | mariadb:10.11 | mariadb:10.6 |
+| --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
+| **8.0** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.1** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.2** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.3** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.4** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+
+### WordPress 6.7
+
+| PHP \ DB | mysql:9 | mysql:8.4 | mysql:8.0 | mariadb:12 | mariadb:11.8 | mariadb:11.4 | mariadb:11.2 | mariadb:11.0 | mariadb:10.11 | mariadb:10.6 |
+| --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
+| **8.0** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.1** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.2** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.3** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.4** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+
+### WordPress 6.8
+
+| PHP \ DB | mysql:9 | mysql:8.4 | mysql:8.0 | mariadb:12 | mariadb:11.8 | mariadb:11.4 | mariadb:11.2 | mariadb:11.0 | mariadb:10.11 | mariadb:10.6 |
+| --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
+| **8.0** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.1** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.2** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.3** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.4** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+
+### WordPress 6.9
+
+| PHP \ DB | mysql:9 | mysql:8.4 | mysql:8.0 | mariadb:12 | mariadb:11.8 | mariadb:11.4 | mariadb:11.2 | mariadb:11.0 | mariadb:10.11 | mariadb:10.6 |
+| --- | --- | --- | --- | --- | --- | --- | --- | --- | --- | --- |
+| **8.0** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.1** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.2** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.3** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+| **8.4** | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ | ✓ |
+
+</details>
+
 ## Change Log ##
 * 1.2.1 - Updated dev dependencies, PHP min to 8.0, WP test matrix to 6.6–6.9, aligned configs with shared standards.
 * 1.2.0 - Added in helper method for unsigned_big() (thanks [iniznet](https://github.com/iniznet)), updated dev deps and GH pipelines.
