+++ tests/Intergration/Test_Table_With_Indexes.php
@@ -21,11 +21,11 @@ use PinkCrab\PHPUnit_Helpers\Reflection;
 use PinkCrab\Table_Builder\Table_Schema;
 use PinkCrab\Table_Builder\Builders\DB_Delta;
 use PinkCrab\Table_Builder\Engines\WPDB_DB_Delta\DB_Delta_Engine;
-use PinkCrab\Table_Builder\Tests\Int_Width_Aware_Trait;
+use PinkCrab\Table_Builder\Tests\DB_Column_Meta_Compat_Trait;
 
 class Test_Table_With_Indexes extends WP_UnitTestCase {
 
-	use Int_Width_Aware_Trait;
+	use DB_Column_Meta_Compat_Trait;
 
 
 
@@ -107,26 +107,26 @@ class Test_Table_With_Indexes extends WP_UnitTestCase {
 				'Null'    => 'NO',
 				'Key'     => 'UNI',
 				'Default' => null,
-				'Extra'   => '',
+				'Extra'   => $this->extra_empty(),
 			),
 			'user_email'   => array(
 				'Type'    => 'varchar(256)',
 				'Null'    => 'NO',
 				'Key'     => 'UNI',
 				'Default' => null,
-				'Extra'   => '',
+				'Extra'   => $this->extra_empty(),
 			),
 			'created_on'   => array(
 				'Type'  => 'datetime',
 				'Null'  => 'NO',
 				'Key'   => '',
-				'Extra' => '',
+				'Extra' => $this->extra_empty(),
 			),
 			'last_updated' => array(
 				'Type'  => 'datetime',
 				'Null'  => 'NO',
 				'Key'   => '',
-				'Extra' => '',
+				'Extra' => $this->extra_empty(),
 			),
 		);
